+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.0-beta.3
+
+- Added Ubuntu Dock support (#1). Thanks to @0xHertz for the report and
+  initial patch.
+- Tested with Ubuntu Dock on Ubuntu 24.04 (GNOME Shell 46).
+
 ## 0.1.0-beta.2
 
 - Added motion on the GNOME overview dash: hover, press, and launch.